+++ kazuki-api/src/main/java/io/kazuki/v0/store/management/ComponentRegistrar.java
@@ -12,23 +12,15 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.kazuki.v0.store.schema;
+package io.kazuki.v0.store.management;
 
-import io.kazuki.v0.store.KazukiException;
-import io.kazuki.v0.store.Version;
-import io.kazuki.v0.store.keyvalue.KeyValuePair;
-import io.kazuki.v0.store.management.KazukiComponent;
-import io.kazuki.v0.store.schema.model.Schema;
-
-public interface SchemaStore extends KazukiComponent<SchemaStore> {
-  KeyValuePair<Schema> retrieveSchema(String type) throws KazukiException;
-
-  Version createSchema(String type, Schema value) throws KazukiException;
-
-  Version updateSchema(final String type, final Version version, final Schema value)
-      throws KazukiException;
-
-  boolean deleteSchema(final String type, final Version version) throws KazukiException;
-
-  void clear() throws KazukiException;
+/**
+ * Registration interface for the "management side". A service implements this class because it
+ * wants to receive registration callbacks from Kazuki components.
+ */
+public interface ComponentRegistrar {
+  /**
+   * Called only once by each Kauzki component to register its descriptor with the manager.
+   */
+  void register(ComponentDescriptor component);
 }