+++ kazuki-impl/src/main/java/io/kazuki/v0/store/index/SecondaryIndexTableHelper.java
@@ -44,6 +44,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+import javax.inject.Inject;
+
 import org.skife.jdbi.v2.Handle;
 import org.skife.jdbi.v2.IDBI;
 import org.skife.jdbi.v2.TransactionCallback;
@@ -51,7 +53,6 @@ import org.skife.jdbi.v2.TransactionStatus;
 import org.skife.jdbi.v2.exceptions.UnableToExecuteStatementException;
 
 import com.google.common.base.Throwables;
-import com.google.inject.Inject;
 
 public class SecondaryIndexTableHelper {
   private final SqlTypeHelper typeHelper;