+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+=======
+v0.5.12
+---
+* initial implementation of KazukiManager component introspection API
+
+v0.5.11
+---
+* removed "requireExplicitBindings()" - hopefully back soon!
+
+v0.5.10
+---
+* fixed KazukiModule "Key" binding
+
 v0.5.9
 ---
 * added parseKey() and parseVersion() to KeyValueStore and SequenceService APIs